python/worker/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
xgboost==2.0.3
scikit-learn==1.5.0
joblib==1.4.2
numpy==1.26.4
azure-storage-blob==12.20.0
